requirements.txt
typer>=0.12.3
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
requests>=2.32.2
pytest>=8.2.0